Added new cities file from OpenWeatherMap and updated cities comparison script in order to process it
Since the OpenWeatherMap limits the number of queries to 60/mn but allow
for bulk queries, we use their json containing 200k+ cities to extract
cities id to make a bulk query later on.

We thus have to run the coordinates comparison script to retain cities
that don't fall on the same voxel.

Warning : it takes a long time to run the script with that much cities
(the algo to find duplicate has been improved by a factor 10 to reduce
the time but the next step is still not optimized)